+++ library/Hasql/Prelude.hs
@@ -1,15 +1,13 @@
 module Hasql.Prelude
 ( 
   module Exports,
-  bug,
-  bottom,
 )
 where
 
 
 -- base-prelude
 -------------------------
-import BasePrelude as Exports
+import BasePrelude as Exports hiding (left, right, isLeft, isRight)
 
 -- transformers
 -------------------------
@@ -19,21 +17,18 @@ import Control.Monad.Trans.Class as Exports
 import Control.Monad.IO.Class as Exports
 import Data.Functor.Identity as Exports
 
+-- mtl
+-------------------------
+import Control.Monad.Error.Class as Exports
+
 -- mmorph
 -------------------------
 import Control.Monad.Morph as Exports
 
--- monad-control
+-- either
 -------------------------
-import Control.Monad.Trans.Control as Exports hiding (embed, embed_)
-
--- transformers-base
--------------------------
-import Control.Monad.Base as Exports
-
--- safe
--------------------------
-import Safe as Exports
+import Control.Monad.Trans.Either as Exports
+import Data.Either.Combinators as Exports
 
 -- list-t
 -------------------------
@@ -46,26 +41,3 @@ import Data.Vector as Exports (Vector)
 -- text
 -------------------------
 import Data.Text as Exports (Text)
-
--- bytestring
--------------------------
-import Data.ByteString as Exports (ByteString)
-
--- time
--------------------------
-import Data.Time as Exports
-
--- placeholders
--------------------------
-import Development.Placeholders as Exports
-
--- custom
--------------------------
-import qualified Debug.Trace.LocationTH
-
-
-bug = [e| $(Debug.Trace.LocationTH.failure) . (msg <>) |]
-  where
-    msg = "A \"hasql\" package bug: " :: String
-
-bottom = [e| $bug "Bottom evaluated" |]